+++ src/photon_platform/clerk/modulator/__init__.py
@@ -11,8 +11,3 @@ tools for creating python modules and their components
 
 from .modulator import *
 
-__author__ = "phiarchitect"
-__maintainer__ = "PHOTON platform"
-__email__ = "github@phiarchitect.com"
-__version__ = "0.0.1"
-__licence__ = "MIT"