+++ src/photon_platform/clerk/progenitor/__init__.py
@@ -11,8 +11,3 @@ tools for creating and initializing Python projects
 
 from .progenitor import create_project
 
-__author__ = "phiarchitect"
-__maintainer__ = "PHOTON platform"
-__email__ = "github@phiarchitect.com"
-__version__ = "0.0.1"
-__licence__ = "MIT"